Cleaning files headers

+++ src/Roadiz/Core/Services/TwigServiceProvider.php
@@ -249,8 +249,8 @@ class TwigServiceProvider implements ServiceProviderInterface
                 'centralTruncate',
                 function ($object, $length, $offset = 0, $ellipsis = "[…]") {
                     if (strlen($object) > $length + strlen($ellipsis)) {
-                        $str1 = substr($object, 0, floor($length / 2) + floor($offset / 2));
-                        $str2 = substr($object, (floor($length / 2) * -1) + floor($offset / 2));
+                        $str1 = substr($object, 0, (int) (floor($length / 2) + floor($offset / 2)));
+                        $str2 = substr($object, (int) ((floor($length / 2) * -1) + floor($offset / 2)));
 
                         return $str1 . $ellipsis . $str2;
                     } else {
